Simplified exception

+++ src/Service/AListService.php
@@ -14,7 +14,7 @@ class AListService
     public function __construct($url, $options = [])
     {
         if(empty($options['token'])){
-            throw new \Exception('AList token is required. Please check your configuration.');
+            throw new ApiException('AList token is required. Please check your configuration.');
         }
         $this->client=new AListClient($url,$options['token'],$options['client']??[]);
     }
@@ -23,6 +23,7 @@ class AListService
     }
     public function getDownloadUrl($path,$expiration=0)
     {
+        $path=Path::clean($path);
         return $this->client->getDownloadUrl($path,$expiration);
     }
     function copy($srcDir, $dstDir, $names)
@@ -57,10 +58,12 @@ class AListService
     function read($path)
     {
         $path=Path::clean($path);
+        if(!$this->get($path)){
+            return null;
+        }
         $response=$this->client->fsLink(['path'=>$path]);
         $url=$response['data']['url']??'';
         if($url){
-            //Stream reading from url
             $guzzleClient=$this->client->getGuzzleClient();
             return $guzzleClient->get($url)->getBody()->detach();
         }
@@ -136,18 +139,72 @@ class AListService
          * 'hashinfo' => 'null',
          * 'hash_info' => NULL,
          */
-        $visibility = StorageContract::VISIBILITY_PRIVATE;
+        $visibility = StorageContract::VISIBILITY_PUBLIC;
         return [
-            StorageAttributes::ATTRIBUTE_PATH => ltrim($path,'\/'),
+            StorageAttributes::ATTRIBUTE_PATH => $path,
             StorageAttributes::ATTRIBUTE_LAST_MODIFIED => strtotime($file['modified']),
             StorageAttributes::ATTRIBUTE_FILE_SIZE => $file['size'],
             StorageAttributes::ATTRIBUTE_TYPE => $file['is_dir'] ? 'dir' : 'file',
-            StorageAttributes::ATTRIBUTE_MIME_TYPE => $response['type'] ?? null,
+            StorageAttributes::ATTRIBUTE_MIME_TYPE => $this->guessMimeTypeFromFileName($path),
             StorageAttributes::ATTRIBUTE_VISIBILITY=>$visibility,
             '@id'=>'',
             '@link' => '',
             '@shareLink'=>'',
             '@downloadUrl' => '',
         ];
+    }
+    protected function guessMimeTypeFromFileName($filename)
+    {
+        $ext = pathinfo($filename, PATHINFO_EXTENSION);
+        //Use predefined mime types
+        $mimes=[
+            'txt'=>'text/plain',
+            'html'=>'text/html',
+            'htm'=>'text/html',
+            'php'=>'text/html',
+            'css'=>'text/css',
+            'js'=>'application/javascript',
+            'json'=>'application/json',
+            'xml'=>'application/xml',
+            'swf'=>'application/x-shockwave-flash',
+            'flv'=>'video/x-flv',
+            //Images
+            'png'=>'image/png',
+            'jpe'=>'image/jpeg',
+            'jpeg'=>'image/jpeg',
+            'jpg'=>'image/jpeg',
+            'gif'=>'image/gif',
+            'bmp'=>'image/bmp',
+            'ico'=>'image/vnd.microsoft.icon',
+            'tiff'=>'image/tiff',
+            'tif'=>'image/tiff',
+            'svg'=>'image/svg+xml',
+            'svgz'=>'image/svg+xml',
+            //Archives
+            'zip'=>'application/zip',
+            'rar'=>'application/x-rar-compressed',
+            'exe'=>'application/x-msdownload',
+            'msi'=>'application/x-msdownload',
+            'cab'=>'application/vnd.ms-cab-compressed',
+            //Audio/video
+            'mp3'=>'audio/mpeg',
+            'qt'=>'video/quicktime',
+            'mov'=>'video/quicktime',
+            //Adobe
+            'pdf'=>'application/pdf',
+            'psd'=>'image/vnd.adobe.photoshop',
+            'ai'=>'application/postscript',
+            'eps'=>'application/postscript',
+            'ps'=>'application/postscript',
+            //MS Office
+            'doc'=>'application/msword',
+            'rtf'=>'application/rtf',
+            'xls'=>'application/vnd.ms-excel',
+            'ppt'=>'application/vnd.ms-powerpoint',
+            //Open Office
+            'odt'=>'application/vnd.oasis.opendocument.text',
+            'ods'=>'application/vnd.oasis.opendocument.spreadsheet',
+        ];
+        return $mimes[$ext]??'';
     }
 }
